+++ app/app/src/main/java/hk/edu/polyu/comp/hellohongkong/app/EventsInfo.java
@@ -64,8 +64,8 @@ public class EventsInfo extends AppCompatActivity {
         for (Event lvEvent : lvEventList) {
             HashMap<String, String> lvItem = new HashMap<>();
             lvItem.put("name", lvEvent.getName());
-            lvItem.put("start", Integer.toString(lvEvent.getStartTime()));
-            lvItem.put("end", Integer.toString(lvEvent.getEndTime()));
+            lvItem.put("start", Long.toString(lvEvent.getStartTime()));
+            lvItem.put("end", Long.toString(lvEvent.getEndTime()));
             lvItem.put("location", lvEvent.getLocation());
             lvItem.put("description", lvEvent.getDescription());
             lvArrayList.add(lvItem);
